examples/auth-peewee/requirements.txt
Flask
Flask-Admin
#Flask-SQLAlchemy
Flask-Peewee
Flask-Security>=1.7.5
